eleventh homework

+++ tp12_mvc2/app/model/ModelVin.php
@@ -159,16 +159,25 @@ class ModelVin
     {
         try {
             $database = Model::getInstance();
-
-            $query = "DELETE FROM vin WHERE id = :id";
-            $statement = $database->prepare($query);
-            $statement->execute([
+            $query = "SELECT * FROM recolte WHERE vin_id = :id";
+            $testExistence = $database->prepare($query);
+            $testExistence->execute([
                     'id' => $id
             ]);
-            return $id;
+            if ($testExistence->rowCount() > 0) {
+                return null;
+            } else {
+                $results = ModelVin::getOne($id);
+                $query = "DELETE FROM vin WHERE id = :id";
+                $statement = $database->prepare($query);
+                $statement->execute([
+                        'id' => $id
+                ]);
+                return $results[0];
+            }
         } catch (PDOException $e) {
             printf("%s - %s<p/>\n", $e->getCode(), $e->getMessage());
-            return -1;
+            return -1; // Error exit code
         }
     }
